' ' Add Hands free Capture Modes:
- Voice Shutter: Voice triggered snapshots
If one of the pre-specified hotwords is detected, fire the shutter
and take a picture

Change-Id: I8fc716c836728f82fd69741a6c1c7864d86934c4

- Camera : Timer Snapshot
Allow user to take a snapshot after a certain amount of time.
Credits : Michael Simon for the base port
	  David Van Tonder for indicator icon

Change-Id: Ic50bf1c1f7fd85d8c48dac3e51766aeff2b381ea

- Merge voice and timer options into a single "no-hands shutter" option

... and add a first-use hint to VS listing the hotwords
Usage of handsfree modes can be made continuous by flipping a switch
in the settings pane. The default behavior is to reset the shutter
mode to normal after a picture is taken

Change-Id: I271c7ba71f320e19be5d1a6a8d98dafeb4b8385a

Additional drawables included from Android system due to size inconsistencies

thanks to:rmcc
commit 0bb5084'